ajustes oficiales de servidor

+++ app/Http/Controllers/BlogController.php
@@ -15,6 +15,7 @@ use App\User;
 use App\comentario;
 use App\config;
 
+
 class BlogController extends Controller {
 
     /**
@@ -22,14 +23,18 @@ class BlogController extends Controller {
      *
      * @return \Illuminate\Http\Response
      */
-    public function gestion() {
+    public function gestion() {        
+        if(\Auth::user() && (\Auth::user()->role == 'editor' || \Auth::user()->role == 'socio' || \Auth::user()->role == 'admin' || \Auth::user()->role == 'Superadmin')){
 
-        $blog = blog::all();
+            $blog = blog::all();
 
 
-        return view('blog.gestion', [
-            'blogs' => $blog
-        ]);
+            return view('blog.gestion', [
+                'blogs' => $blog
+            ]);
+        }else{
+            return redirect()->route('index');
+        }
     }
 
     /**
@@ -47,12 +52,17 @@ class BlogController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function create() {
-        $categorias = categoria::all();
-        $tag = \App\tag::all();
-        return view('blog.crear', [
-            'categorias' => $categorias,
-            'tags' => $tag
-        ]);
+        if(\Auth::user() && (\Auth::user()->role == 'editor' || \Auth::user()->role == 'socio' || \Auth::user()->role == 'admin' || \Auth::user()->role == 'Superadmin')){
+
+            $categorias = categoria::all();
+            $tag = \App\tag::all();
+            return view('blog.crear', [
+                'categorias' => $categorias,
+                'tags' => $tag
+            ]);
+        }else{
+            return redirect()->route('index');
+        }
     }
 
     /**
@@ -62,63 +72,68 @@ class BlogController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request) {
-        //validacion
-        $validate = $this->validate($request, [
-            'categoria' => 'required',
-            'titulo' => 'required',
-            'imagen' => 'required|image',
-            'subtitulo' => 'required',
-            'contenido' => 'required',
-            'tag' => 'required',
-            'metadesc' => 'required',
-            'metatitle' => 'required',
-            'metakey' => 'required'
-        ]);
-        //recojo los datos
+        if(\Auth::user() && (\Auth::user()->role == 'editor' || \Auth::user()->role == 'socio' || \Auth::user()->role == 'admin' || \Auth::user()->role == 'Superadmin')){
 
-
-        $categoria = $request->input('categoria');
-        $titulo = $request->input('titulo');
-        $imagen = $request->file('imagen');
-        $subtitulo = $request->input('subtitulo');
-        $contenido = $request->input('contenido');
-        $tag = $request->input('tag');
-        $metadesc = $request->input('metadesc');
-        $metatitle = $request->input('metatitle');
-        $metakey = $request->input('metakey');
-
-        //seteo objeto
-        //cargo usuario autenticado
-        $user = \Auth::user();
-        $blog = new blog();
-        $blog->user_id = $user->id;
-        $blog->categoria_id = $categoria;
-        $blog->titulo = $titulo;
-        if ($imagen) {
-            //nombre unico
-            $imagen_n = time() . $imagen->getClientOriginalName();
-            //guardamos en la carpeta
-            Storage::disk('blog')->put($imagen_n, File::get($imagen));
-            //seteo el objeto
-            $blog->imagen = $imagen_n;
-        }
-        $blog->subtitulo = $subtitulo;
-        $blog->contenido = $contenido;
-        $blog->tag_id = $tag;
-        $blog->metadesc = $metadesc;
-        $blog->metatitle = $metatitle;
-        $blog->metakey = $metakey;
-        //metodo save
-        $save = $blog->save();
-        //comprobamos redireccion
-        if ($save) {
-            return redirect()->route('blog.gestion')->with([
-                        'message' => 'Articulo Creado Exitosamente'
+            //validacion
+            $validate = $this->validate($request, [
+                'categoria' => 'required',
+                'titulo' => 'required',
+                'imagen' => 'required|image',
+                'subtitulo' => 'required',
+                'contenido' => 'required',
+                'tag' => 'required',
+                'metadesc' => 'required',
+                'metatitle' => 'required',
+                'metakey' => 'required'
             ]);
-        } else {
-            return redirect()->route('blog.gestion')->with([
-                        'message' => 'Fallo Al Crear El Post'
-            ]);
+            //recojo los datos
+
+
+            $categoria = $request->input('categoria');
+            $titulo = $request->input('titulo');
+            $imagen = $request->file('imagen');
+            $subtitulo = $request->input('subtitulo');
+            $contenido = $request->input('contenido');
+            $tag = $request->input('tag');
+            $metadesc = $request->input('metadesc');
+            $metatitle = $request->input('metatitle');
+            $metakey = $request->input('metakey');
+
+            //seteo objeto
+            //cargo usuario autenticado
+            $user = \Auth::user();
+            $blog = new blog();
+            $blog->user_id = $user->id;
+            $blog->categoria_id = $categoria;
+            $blog->titulo = $titulo;
+            if ($imagen) {
+                //nombre unico
+                $imagen_n = time() . $imagen->getClientOriginalName();
+                //guardamos en la carpeta
+                Storage::disk('blog')->put($imagen_n, File::get($imagen));
+                //seteo el objeto
+                $blog->imagen = $imagen_n;
+            }
+            $blog->subtitulo = $subtitulo;
+            $blog->contenido = $contenido;
+            $blog->tag_id = $tag;
+            $blog->metadesc = $metadesc;
+            $blog->metatitle = $metatitle;
+            $blog->metakey = $metakey;
+            //metodo save
+            $save = $blog->save();
+            //comprobamos redireccion
+            if ($save) {
+                return redirect()->route('blog.gestion')->with([
+                            'message' => 'Articulo Creado Exitosamente'
+                ]);
+            } else {
+                return redirect()->route('blog.gestion')->with([
+                            'message' => 'Fallo Al Crear El Post'
+                ]);
+            }
+        }else{
+            return redirect()->route('index');
         }
     }
 
@@ -157,14 +172,19 @@ class BlogController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function edit(blog $blog, $id) {
-        $blog = blog::find($id);
-        $categoria = categoria::all();
-        $tag = \App\tag::all();
-        return view('blog.editar', [
-            'blog' => $blog,
-            'categorias' => $categoria,
-            'tags' => $tag
-        ]);
+        if(\Auth::user() && (\Auth::user()->role == 'editor' || \Auth::user()->role == 'socio' || \Auth::user()->role == 'admin' || \Auth::user()->role == 'Superadmin')){
+
+            $blog = blog::find($id);
+            $categoria = categoria::all();
+            $tag = \App\tag::all();
+            return view('blog.editar', [
+                'blog' => $blog,
+                'categorias' => $categoria,
+                'tags' => $tag
+            ]);
+        }else{
+            return redirect()->route('index');
+        }
     }
 
     /**
@@ -175,56 +195,61 @@ class BlogController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, blog $blog) {
-        //validar formulario
-        $validate = $this->validate($request, [
-            'imagen' => 'image'
-        ]);
-        //guardar datos
-        $id = $request->input('id');
-        $categoria = $request->input('categoria');
+        if(\Auth::user() && (\Auth::user()->role == 'editor' || \Auth::user()->role == 'socio' || \Auth::user()->role == 'admin' || \Auth::user()->role == 'Superadmin')){
 
-        $titulo = $request->input('titulo');
-        $imagen = $request->file('imagen');
-        $subtitulo = $request->input('subtitulo');
-        $contenido = $request->input('contenido');
-        $tag = $request->input('tag');
+            //validar formulario
+            $validate = $this->validate($request, [
+                'imagen' => 'image'
+            ]);
+            //guardar datos
+            $id = $request->input('id');
+            $categoria = $request->input('categoria');
 
-        $metadesc = $request->input('metadesc');
-        $metatitle = $request->input('metatitle');
-        $metakey = $request->input('metakey');
-        //cargo objeto de la BD
-        $user = \Auth::user();
-        $blog = blog::find($id);
-        //setear objetos
-        $blog->user_id = $user->id;
-        $blog->categoria_id = $categoria;
-        $blog->titulo = $titulo;
-        //recibir imagen
-        if ($imagen) {
-            //poner nombre unico
-            $imagen_name = time() . $imagen->getClientOriginalName();
-            //guardamos en la carpeta blogs de storage/app/blogs
-            Storage::disk('blogs')->put($imagen_name, File::get($imagen));
-            //seteo imagen en el objeto
-            $blog->imagen = $imagen_name;
-        }
-        $blog->subtitulo = $subtitulo;
-        $blog->contenido = $contenido;
-        $blog->tag_id = $tag;
-        $blog->metadesc = $metadesc;
-        $blog->metatitle = $metatitle;
-        $blog->metakey = $metakey;
-        //guardar
-        $save = $blog->update();
-        //redirigir
-        if ($save) {
-            return redirect()->route('blog.gestion')->with([
-                        'message' => 'blog guardado exitosamente'
-            ]);
-        } else {
-            return redirect()->route('blog.gestion')->with([
-                        'message' => 'fallo al guardar blog'
-            ]);
+            $titulo = $request->input('titulo');
+            $imagen = $request->file('imagen');
+            $subtitulo = $request->input('subtitulo');
+            $contenido = $request->input('contenido');
+            $tag = $request->input('tag');
+
+            $metadesc = $request->input('metadesc');
+            $metatitle = $request->input('metatitle');
+            $metakey = $request->input('metakey');
+            //cargo objeto de la BD
+            $user = \Auth::user();
+            $blog = blog::find($id);
+            //setear objetos
+            $blog->user_id = $user->id;
+            $blog->categoria_id = $categoria;
+            $blog->titulo = $titulo;
+            //recibir imagen
+            if ($imagen) {
+                //poner nombre unico
+                $imagen_name = time() . $imagen->getClientOriginalName();
+                //guardamos en la carpeta blogs de storage/app/blogs
+                Storage::disk('blogs')->put($imagen_name, File::get($imagen));
+                //seteo imagen en el objeto
+                $blog->imagen = $imagen_name;
+            }
+            $blog->subtitulo = $subtitulo;
+            $blog->contenido = $contenido;
+            $blog->tag_id = $tag;
+            $blog->metadesc = $metadesc;
+            $blog->metatitle = $metatitle;
+            $blog->metakey = $metakey;
+            //guardar
+            $save = $blog->update();
+            //redirigir
+            if ($save) {
+                return redirect()->route('blog.gestion')->with([
+                            'message' => 'blog guardado exitosamente'
+                ]);
+            } else {
+                return redirect()->route('blog.gestion')->with([
+                            'message' => 'fallo al guardar blog'
+                ]);
+            }
+        }else{
+            return redirect()->route('index');
         }
     }
 
@@ -235,17 +260,22 @@ class BlogController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy($id) {
-        $blog = blog::find($id);
-        //eliminar registro
-        $save = $blog->delete();
-        if ($save) {
-            return redirect()->route('blog.gestion')->with([
-                        'message' => 'blog eliminado exitosamente'
-            ]);
-        } else {
-            return redirect()->route('blog.gestion')->with([
-                        'message' => 'fallo al eliminar blog'
-            ]);
+            if(\Auth::user() && (\Auth::user()->role == 'socio' || \Auth::user()->role == 'admin' || \Auth::user()->role == 'Superadmin')){
+
+            $blog = blog::find($id);
+            //eliminar registro
+            $save = $blog->delete();
+            if ($save) {
+                return redirect()->route('blog.gestion')->with([
+                            'message' => 'blog eliminado exitosamente'
+                ]);
+            } else {
+                return redirect()->route('blog.gestion')->with([
+                            'message' => 'fallo al eliminar blog'
+                ]);
+            }
+        }else{
+            return redirect()->route('index');
         }
     }
